contracts/ens: moved cid decoding to own file
