+++ contracts/ens/ens_test.go
@@ -17,10 +17,6 @@
 package ens
 
 import (
-	"bytes"
-	"encoding/binary"
-	"encoding/hex"
-	"fmt"
 	"math/big"
 	"testing"
 
@@ -124,169 +120,3 @@ func TestENS(t *testing.T) {
 	}
 	t.Fatal("todo: try to set old contract with new multicodec stuff and assert fail, set new contract with multicodec stuff, encode, decode and assert returns correct hash")
 }
-
-func TestEIPSpecCidDecode(t *testing.T) {
-	/*storage system: IPFS (0xe3)
-	  CID version: 1 (0x01)
-	  content type: dag-pb (0x70)
-	  hash function: sha2-256 (0x12)
-	  hash length: 32 bytes (0x20)
-	  hash: 29f2d17be6139079dc48696d1f582a8530eb9805b561eda517e22a892c7e3f1f
-	*/
-
-	const eipSpecHash = "e3010170122029f2d17be6139079dc48696d1f582a8530eb9805b561eda517e22a892c7e3f1f"
-	const eipHash = "29f2d17be6139079dc48696d1f582a8530eb9805b561eda517e22a892c7e3f1f"
-
-	b, err := hex.DecodeString(eipSpecHash)
-	if err != nil {
-		t.Fatal(err)
-	}
-	hashBytes, err := hex.DecodeString(eipHash)
-
-	if err != nil {
-		t.Fatal(err)
-	}
-	h, err := manualDecode(b)
-
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	if !bytes.Equal(h[:], hashBytes) {
-		t.Fatal("should be equal")
-	}
-
-}
-
-func TestManualCidDecode(t *testing.T) {
-	// call cid encode method with hash. expect byte slice returned, compare according to spec
-	bb := []byte{}
-
-	for _, v := range []struct {
-		name        string
-		headerBytes []byte
-		fails       bool
-	}{
-		{
-			name:        "values correct, should not fail",
-			headerBytes: []byte{0xe4, 0x01, 0x99, 0x1b, 0x20},
-			fails:       false,
-		},
-		{
-			name:        "cid version wrong, should fail",
-			headerBytes: []byte{0xe4, 0x01, 0x99, 0x1b, 0x20},
-			fails:       true,
-		},
-		{
-			name:        "hash length wrong, should fail",
-			headerBytes: []byte{0xe4, 0x01, 0x99, 0x1b, 0x1f},
-			fails:       true,
-		},
-		{
-			name:        "values correct for ipfs, should fail",
-			headerBytes: []byte{0xe3, 0x01, 0x99, 0x1b, 0x20},
-			fails:       true,
-		},
-	} {
-		t.Run(v.name, func(t *testing.T) {
-			buf := make([]byte, binary.MaxVarintLen64)
-			for _, vv := range v.headerBytes {
-				n := binary.PutUvarint(buf, uint64(vv))
-				bb = append(bb, buf[:n]...)
-			}
-
-			h := common.HexToHash("29f2d17be6139079dc48696d1f582a8530eb9805b561eda517e22a892c7e3f1f")
-			bb = append(bb, h[:]...)
-			str := hex.EncodeToString(bb)
-			fmt.Println(str)
-			decodedHash, e := manualDecode(bb)
-			switch v.fails {
-			case true:
-				if e == nil {
-					t.Fatal("the decode should fail")
-				}
-			case false:
-				if e != nil {
-					t.Fatal("the deccode shouldnt fail")
-				}
-			}
-			if e != nil {
-				t.Fatal(e)
-			}
-
-			if !bytes.Equal(decodedHash[:], h[:]) {
-				t.Fatal("hashes not equal")
-			}
-		})
-	}
-
-	/* from the EIP documentation
-	   storage system: Swarm (0xe4)
-	   CID version: 1 (0x01)
-	   content type: swarm-manifest (0x??)
-	   hash function: keccak-256 (0x1B)
-	   hash length: 32 bytes (0x20)
-	   hash: 29f2d17be6139079dc48696d1f582a8530eb9805b561eda517e22a892c7e3f1f
-	*/
-
-}
-
-func TestManuelCidEncode(t *testing.T) {
-	// call cid encode method with hash. expect byte slice returned, compare according to spec
-
-	/* from the EIP documentation
-	   storage system: Swarm (0xe4)
-	   CID version: 1 (0x01)
-	   content type: swarm-manifest (0x??)
-	   hash function: keccak-256 (0x1B)
-	   hash length: 32 bytes (0x20)
-	   hash: 29f2d17be6139079dc48696d1f582a8530eb9805b561eda517e22a892c7e3f1f
-	*/
-
-}
-
-/*
-func TestCIDSanity(t *testing.T) {
-	hashStr := "d1de9994b4d039f6548d191eb26786769f580809256b4685ef316805265ea162"
-	hash := common.HexToHash(hashStr) //this always yields a 32 byte long hash
-	cc, err := encodeCid(hash)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	if cc.Prefix().MhLength != 32 {
-		t.Fatal("w00t")
-	}
-	decoded, err := mh.Decode(cc.Hash())
-	if err != nil {
-		t.Fatal(err)
-	}
-	if decoded.Length != 32 {
-		t.Fatal("invalid length")
-	}
-	if !bytes.Equal(decoded.Digest, hash[:]) {
-		t.Fatalf("hashes not equal")
-	}
-
-	if decoded.Length != 32 {
-		t.Fatal("wrong length")
-	}
-	fmt.Println(cc.StringOfBase(multibase.Base16))
-
-	bbbb, e := cc.StringOfBase(multibase.Base16)
-	if e != nil {
-		t.Fatal(e)
-	}
-	fmt.Println(bbbb)
-	//create the CID string artificially
-	hashStr = "f01551b20" + hashStr
-
-	c, err := cid.Decode(hashStr)
-	if err != nil {
-		t.Fatalf("Error decoding CID: %v", err)
-	}
-
-	fmt.Sprintf("Got CID: %v", c)
-	fmt.Println("Got CID:", c.Prefix())
-
-}*/
